Move common types to own folder
