+++ wiki/src/Layout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Navigation from './Navigation';
 import logo from './images/logo.png';
-import ContentTree from './commons/ContentTree';
+import ContentTree from './commons/types/ContentTree';
 import BusyMask from './commons/BusyMask';
  
 interface ILayoutProps {
