+++ wiki/src/Navigation.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import TreeView from './commons/tree-view/TreeView';
-import ContentTree from './commons/ContentTree';
+import ContentTree from './commons/types/ContentTree';
 
 interface INavigationProps {
   contentTree: ContentTree;
